WIP: Partial refactoring in order to follow hexagonal principles
