+++ project/src/main/java/es/ubu/lsi/avrela/apm/adapter/github/model/GitHubComment.java
@@ -1,4 +1,4 @@
-package es.ubu.lsi.avrela.client.github;
+package es.ubu.lsi.avrela.apm.adapter.github.model;
 
 import com.google.gson.annotations.SerializedName;
 import java.time.ZonedDateTime;
