+++ project/src/main/java/es/ubu/lsi/avrela/apm/adapter/github/model/GitHubMilestone.java
@@ -1,6 +1,7 @@
-package es.ubu.lsi.avrela.client.github;
+package es.ubu.lsi.avrela.apm.adapter.github.model;
 
 import com.google.gson.annotations.SerializedName;
+import es.ubu.lsi.avrela.apm.adapter.github.model.GitHubItemState;
 import java.time.ZonedDateTime;
 import lombok.Data;
