+++ project/src/main/java/es/ubu/lsi/avrela/apm/adapter/github/model/GitHubMilestoneState.java
@@ -1,0 +1,6 @@
+package es.ubu.lsi.avrela.apm.adapter.github.model;
+
+public enum GitHubMilestoneState {
+
+
+}
