+++ project/src/main/java/es/ubu/lsi/avrela/apm/domain/model/Sprint.java
@@ -1,7 +1,6 @@
-package es.ubu.lsi.avrela.domain;
+package es.ubu.lsi.avrela.apm.domain.model;
 
 import java.util.List;
-import java.util.Set;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
