+++ project/src/main/java/es/ubu/lsi/avrela/apm/domain/model/SprintState.java
@@ -1,4 +1,4 @@
-package es.ubu.lsi.avrela.domain;
+package es.ubu.lsi.avrela.apm.domain.model;
 
 public enum SprintState {
