+++ project/src/main/java/es/ubu/lsi/avrela/apm/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Agile Project Management hexagon.
+ */
+package es.ubu.lsi.avrela.apm;
