--- project/src/main/java/es/ubu/lsi/avrela/client/github/GitHubMilestoneState.java
@@ -1,6 +1,0 @@
-package es.ubu.lsi.avrela.client.github;
-
-public enum GitHubMilestoneState {
-
-
-}
