+++ project/src/main/java/es/ubu/lsi/avrela/domain/Repository.java
@@ -1,5 +1,8 @@
 package es.ubu.lsi.avrela.domain;
 
+import es.ubu.lsi.avrela.apm.domain.model.Issue;
+import es.ubu.lsi.avrela.apm.domain.model.Sprint;
+import es.ubu.lsi.avrela.scm.domain.model.Commit;
 import java.util.List;
 
 /** Models a repository. */
