+++ project/src/main/java/es/ubu/lsi/avrela/scm/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Source Control Management hexagon.
+ */
+package es.ubu.lsi.avrela.scm;
